+++ src/main/java/com/plaid/client/PlaidApiService.java
@@ -3,10 +3,13 @@ package com.plaid.client;
 import com.plaid.client.request.AccountsBalanceGetRequest;
 import com.plaid.client.request.AccountsGetRequest;
 import com.plaid.client.request.AssetReportAuditCopyCreateRequest;
+import com.plaid.client.request.AssetReportAuditCopyGetRequest;
 import com.plaid.client.request.AssetReportAuditCopyRemoveRequest;
 import com.plaid.client.request.AssetReportCreateRequest;
+import com.plaid.client.request.AssetReportFilterRequest;
 import com.plaid.client.request.AssetReportGetRequest;
 import com.plaid.client.request.AssetReportPdfGetRequest;
+import com.plaid.client.request.AssetReportRefreshRequest;
 import com.plaid.client.request.AssetReportRemoveRequest;
 import com.plaid.client.request.AuthGetRequest;
 import com.plaid.client.request.CategoriesGetRequest;
@@ -19,19 +22,14 @@ import com.plaid.client.request.InstitutionsSearchRequest;
 import com.plaid.client.request.ItemAccessTokenInvalidateRequest;
 import com.plaid.client.request.ItemAccessTokenUpdateVersionRequest;
 import com.plaid.client.request.ItemApexProcessorTokenCreateRequest;
-import com.plaid.client.request.ItemCreateRequest;
-import com.plaid.client.request.ItemCredentialsEncryptRequest;
-import com.plaid.client.request.ItemCredentialsUpdateRequest;
-import com.plaid.client.request.ItemDeleteRequest;
 import com.plaid.client.request.ItemDwollaProcessorTokenCreateRequest;
 import com.plaid.client.request.ItemGetRequest;
-import com.plaid.client.request.ItemMfaEncryptRequest;
-import com.plaid.client.request.ItemMfaRequest;
 import com.plaid.client.request.ItemPublicTokenCreateRequest;
 import com.plaid.client.request.ItemPublicTokenExchangeRequest;
 import com.plaid.client.request.ItemRemoveRequest;
 import com.plaid.client.request.ItemStripeTokenCreateRequest;
 import com.plaid.client.request.ItemWebhookUpdateRequest;
+import com.plaid.client.request.SandboxItemFireWebhookRequest;
 import com.plaid.client.request.SandboxItemResetLoginRequest;
 import com.plaid.client.request.SandboxPublicTokenCreateRequest;
 import com.plaid.client.request.TransactionsGetRequest;
@@ -53,25 +51,19 @@ import com.plaid.client.response.InstitutionsSearchResponse;
 import com.plaid.client.response.ItemAccessTokenInvalidateResponse;
 import com.plaid.client.response.ItemAccessTokenUpdateVersionResponse;
 import com.plaid.client.response.ItemApexProcessorTokenCreateResponse;
-import com.plaid.client.response.ItemCreateResponse;
-import com.plaid.client.response.ItemCredentialsEncryptResponse;
-import com.plaid.client.response.ItemCredentialsUpdateResponse;
-import com.plaid.client.response.ItemDeleteResponse;
 import com.plaid.client.response.ItemDwollaProcessorTokenCreateResponse;
 import com.plaid.client.response.ItemGetResponse;
-import com.plaid.client.response.ItemMfaEncryptResponse;
-import com.plaid.client.response.ItemMfaResponse;
 import com.plaid.client.response.ItemPublicTokenCreateResponse;
 import com.plaid.client.response.ItemPublicTokenExchangeResponse;
 import com.plaid.client.response.ItemRemoveResponse;
 import com.plaid.client.response.ItemStripeTokenCreateResponse;
 import com.plaid.client.response.ItemWebhookUpdateResponse;
+import com.plaid.client.response.SandboxItemFireWebhookResponse;
 import com.plaid.client.response.SandboxItemResetLoginResponse;
 import com.plaid.client.response.SandboxPublicTokenCreateResponse;
 import com.plaid.client.response.TransactionsGetResponse;
 import okhttp3.ResponseBody;
 import retrofit2.Call;
-import retrofit2.Response;
 import retrofit2.http.Body;
 import retrofit2.http.POST;
 
@@ -80,23 +72,8 @@ public interface PlaidApiService {
   // item, link, and credentials calls
   ////////////////////////////////////////////////////////
   ////////////////////////////////////////////////////////
-  @POST("/item/create")
-  Call<ItemCreateResponse> itemCreate(@Body ItemCreateRequest request);
-
-  @POST("/item/mfa")
-  Call<ItemMfaResponse> itemMfa(@Body ItemMfaRequest request);
-
   @POST("/item/get")
   Call<ItemGetResponse> itemGet(@Body ItemGetRequest request);
-
-  @POST("/item/credentials/update")
-  Call<ItemCredentialsUpdateResponse> itemCredentialsUpdate(@Body ItemCredentialsUpdateRequest request);
-
-  @POST("/item/credentials/encrypt")
-  Call<ItemCredentialsEncryptResponse> itemCredentialsEncrypt(@Body ItemCredentialsEncryptRequest request);
-
-  @POST("/item/mfa/encrypt")
-  Call<ItemMfaEncryptResponse> itemMfaEncrypt(@Body ItemMfaEncryptRequest request);
 
   @POST("/item/public_token/exchange")
   Call<ItemPublicTokenExchangeResponse> itemPublicTokenExchange(@Body ItemPublicTokenExchangeRequest request);
@@ -119,9 +96,6 @@ public interface PlaidApiService {
   @POST("/item/access_token/update_version")
   Call<ItemAccessTokenUpdateVersionResponse> itemAccessTokenUpdateVersion(@Body ItemAccessTokenUpdateVersionRequest request);
 
-  @POST("/item/delete")
-  Call<ItemDeleteResponse> itemDelete(@Body ItemDeleteRequest request);
-
   @POST("/item/remove")
   Call<ItemRemoveResponse> itemRemove(@Body ItemRemoveRequest request);
 
@@ -130,7 +104,10 @@ public interface PlaidApiService {
 
   // sandbox-only endpoints
   ////////////////////////////////////////////////////////
-  ////////////////////////////////////////////////////////  
+  ////////////////////////////////////////////////////////
+  @POST("/sandbox/item/fire_webhook")
+  Call<SandboxItemFireWebhookResponse> sandboxItemFireWebhook(@Body SandboxItemFireWebhookRequest request);
+
   @POST("/sandbox/item/reset_login")
   Call<SandboxItemResetLoginResponse> sandboxItemResetLogin(@Body SandboxItemResetLoginRequest request);
 
@@ -164,6 +141,15 @@ public interface PlaidApiService {
 
   @POST("/asset_report/audit_copy/remove")
   Call<AssetReportAuditCopyRemoveResponse> assetReportAuditCopyRemove(@Body AssetReportAuditCopyRemoveRequest request);
+
+  @POST("/asset_report/audit_copy/get")
+  Call<AssetReportGetResponse> assetReportAuditCopyGet(@Body AssetReportAuditCopyGetRequest request);
+
+  @POST("/asset_report/refresh")
+  Call<AssetReportCreateResponse> assetReportRefresh(@Body AssetReportRefreshRequest assetReportRefreshRequest);
+
+  @POST("/asset_report/filter")
+  Call<AssetReportCreateResponse> assetReportFilter(@Body AssetReportFilterRequest assetReportFilterRequest);
 
   @POST("/auth/get")
   Call<AuthGetResponse> authGet(@Body AuthGetRequest request);