+++ src/main/java/com/plaid/client/request/InstitutionsGetByIdRequest.java
@@ -1,7 +1,13 @@
 package com.plaid.client.request;
 
+import com.plaid.client.internal.gson.RequiredField;
 import com.plaid.client.request.common.BasePublicRequest;
+import com.plaid.client.request.common.Product;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
+import static com.plaid.client.internal.Util.notEmpty;
 import static com.plaid.client.internal.Util.notNull;
 
 /**
@@ -11,10 +17,32 @@ import static com.plaid.client.internal.Util.notNull;
  */
 public final class InstitutionsGetByIdRequest extends BasePublicRequest {
   private String institutionId;
+  private Options options;
 
   public InstitutionsGetByIdRequest(String institutionId) {
     notNull(institutionId, "institutionId");
 
     this.institutionId = institutionId;
+  }
+
+  public InstitutionsGetByIdRequest withIncludeOptionalMetadata(boolean includeOptionalMetadata) {
+    if (this.options == null) {
+      this.options = new Options();
+    }
+    this.options.includeOptionalMetadata = includeOptionalMetadata;
+    return this;
+  }
+
+  public InstitutionsGetByIdRequest withIncludeStatus(boolean includeStatus) {
+    if (this.options == null) {
+      this.options = new Options();
+    }
+    this.options.includeStatus = includeStatus;
+    return this;
+  }
+
+  private static class Options {
+    private boolean includeOptionalMetadata;
+    private boolean includeStatus;
   }
 }