+++ src/main/java/com/plaid/client/request/InstitutionsSearchRequest.java
@@ -19,7 +19,7 @@ import static com.plaid.client.internal.Util.notNull;
 public final class InstitutionsSearchRequest extends BasePublicRequest {
   private String query;
   private RequiredField<List<Product>> products = RequiredField.empty();
-
+  private Options options;
 
   public InstitutionsSearchRequest(String query) {
     notNull(query, "query");
@@ -35,4 +35,24 @@ public final class InstitutionsSearchRequest extends BasePublicRequest {
     return this;
   }
 
+  public InstitutionsSearchRequest withIncludeOptionalMetadata(boolean includeOptionalMetadata) {
+    if (this.options == null) {
+      this.options = new Options();
+    }
+    this.options.includeOptionalMetadata = includeOptionalMetadata;
+    return this;
+  }
+
+  public InstitutionsSearchRequest withCountryCodes(List<String> countryCodes) {
+    if (this.options == null) {
+      this.options = new Options();
+    }
+    this.options.countryCodes = countryCodes;
+    return this;
+  }
+
+  private static class Options {
+    private boolean includeOptionalMetadata;
+    private List<String> countryCodes;
+  }
 }