+++ src/main/java/com/plaid/client/response/SandboxItemFireWebhookResponse.java
@@ -1,0 +1,14 @@
+package com.plaid.client.response;
+
+import java.util.Date;
+
+/**
+ * Response from /sandbox/item/fire_webhook endpoint.
+ */
+public final class SandboxItemFireWebhookResponse extends BaseResponse {
+  private boolean webhookFired;
+
+  public boolean getWebhookFired() {
+    return webhookFired;
+  }
+}