+++ src/test/java/com/plaid/client/integration/InstitutionsGetTest.java
@@ -24,6 +24,53 @@ public class InstitutionsGetTest extends AbstractIntegrationTest {
   }
 
   @Test
+  public void testSuccessWithProducts() throws Exception {
+    Response<InstitutionsGetResponse> response =
+    client().service().institutionsGet(new InstitutionsGetRequest(3, 0).
+      withCountryCodes(Arrays.asList("US")).
+      withProducts(Arrays.asList(Product.IDENTITY))).
+    execute();
+
+    assertSuccessResponse(response);
+
+    // check number returned
+    assertEquals(3, response.body().getInstitutions().size());
+  }
+
+  @Test
+  public void testSuccessWithIncludeOptionalMetadataTrue() throws Exception {
+    Response<InstitutionsGetResponse> response =
+        client().service().institutionsGet(new InstitutionsGetRequest(3, 0).withIncludeOptionalMetadata(true)).execute();
+
+    assertSuccessResponse(response);
+
+    // check number returned
+    assertEquals(3, response.body().getInstitutions().size());
+  }
+
+  @Test
+  public void testSuccessWithIncludeOptionalMetadataFalse() throws Exception {
+    Response<InstitutionsGetResponse> response =
+        client().service().institutionsGet(new InstitutionsGetRequest(3, 0).withIncludeOptionalMetadata(false)).execute();
+
+    assertSuccessResponse(response);
+
+    // check number returned
+    assertEquals(3, response.body().getInstitutions().size());
+  }
+
+  @Test
+  public void testSuccessWithCountryCodes() throws Exception {
+    Response<InstitutionsGetResponse> response =
+    client().service().institutionsGet(new InstitutionsGetRequest(3, 0).withCountryCodes(Arrays.asList("US"))).execute();
+
+    assertSuccessResponse(response);
+
+    // check number returned
+    assertEquals(3, response.body().getInstitutions().size());
+  }
+
+  @Test
   public void testRequestValidation() throws Exception {
     try {
       new InstitutionsGetRequest(0, 0);