+++ src/test/java/com/plaid/client/integration/InstitutionsSearchTest.java
@@ -6,7 +6,11 @@ import com.plaid.client.response.InstitutionsSearchResponse;
 import org.junit.Test;
 import retrofit2.Response;
 
+import java.util.Arrays;
+
+import static junit.framework.TestCase.assertNull;
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 
 public class InstitutionsSearchTest extends AbstractIntegrationTest {
   @Test
@@ -18,11 +22,49 @@ public class InstitutionsSearchTest extends AbstractIntegrationTest {
   }
 
   @Test
+  public void testSuccessWithIncludeOptionalMetadataTrue() throws Exception {
+    Response<InstitutionsSearchResponse> response =
+        client().service().institutionsSearch(new InstitutionsSearchRequest("t").withIncludeOptionalMetadata(true)).execute();
+
+    assertSuccessResponse(response);
+
+    InstitutionsSearchResponse institutionsSearchResponse = response.body();
+    assertNotNull(institutionsSearchResponse.getInstitutions().get(0).getUrl());
+    assertNotNull(institutionsSearchResponse.getInstitutions().get(0).getPrimaryColor());
+  }
+
+  @Test
+  public void testSuccessWithIncludeOptionalMetadataFalse() throws Exception {
+    Response<InstitutionsSearchResponse> response =
+        client().service().institutionsSearch(new InstitutionsSearchRequest("t").withIncludeOptionalMetadata(false)).execute();
+
+    assertSuccessResponse(response);
+
+    InstitutionsSearchResponse institutionsSearchResponse = response.body();
+
+    assertNull(institutionsSearchResponse.getInstitutions().get(0).getUrl());
+    assertNull(institutionsSearchResponse.getInstitutions().get(0).getPrimaryColor());
+  }
+
+  @Test
+  public void testSuccessWithCountryCodes() throws Exception {
+    Response<InstitutionsSearchResponse> response =
+        client().service().institutionsSearch(new InstitutionsSearchRequest("t").withCountryCodes(Arrays.asList("US"))).execute();
+
+    assertSuccessResponse(response);
+
+    InstitutionsSearchResponse institutionsSearchResponse = response.body();
+
+    assertEquals(institutionsSearchResponse.getInstitutions().get(0).getCountryCodes(), Arrays.asList("US"));
+  }
+
+  @Test
   public void testNoResults() throws Exception {
     Response<InstitutionsSearchResponse> response =
       client().service().institutionsSearch(new InstitutionsSearchRequest("zebra")).execute();
 
     assertSuccessResponse(response);
     assertEquals(0, response.body().getInstitutions().size());
+
   }
 }