+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+# 5.0.0
+
+- Add support for 2019-05-29 API version
+
+# 4.0.3
+
+- Allow more robust customization of underlying OkHttpClient used by PlaidClient.
+
+# 4.0.2
+
+- Add support for [`/sandbox/item/fire_webhook`][sandbox-item-fire-webhook] endpoint ([#140](https://github.com/plaid/plaid-java/pull/140))
+
+# 4.0.1
+
+- Deprecate direct integration as it is no longer supported
+- Replace `/item/delete` with `/item/remove`
+- Add support for include_institution_data for /institutions/search, /institutions/get, /institutions/get_by_id ([#134](https://github.com/plaid/plaid-java/pull/134))
+
+# 3.0.7
+
+- Fix transaction get request handling. ([#132](https://github.com/plaid/plaid-java/pull/132))
+
+# 3.0.6
+
+- Add support for Asset Reports with Insights ([#130](https://github.com/plaid/plaid-java/pull/130))
+
+# 3.0.5
+
+- Add support for filtering asset reports ([#125](https://github.com/plaid/plaid-java/pull/125))
+- Add support for asset report refresh ([#126](https://github.com/plaid/plaid-java/pull/126))
+- Add support for /asset_report/audit_copy/get ([#127](https://github.com/plaid/plaid-java/pull/127))
+
 # 2.2.2
 
 - Allow `TransactionGetRequest.Options` to be subclassed ([#110](https://github.com/plaid/plaid-java/pull/110))
@@ -41,3 +73,5 @@
 # 2.1.0
 
 - Initial release for [Plaid's updated API](https://blog.plaid.com/improving-our-api)
+
+[sandbox-item-fire-webhook]: https://plaid.com/docs/#firing-webhooks